style: restructure & rename & clean up
