+++ FakeNFT/Application/Views/NavBarView.swift
