+++ FakeNFT/Application/Views/RatingView.swift
