+++ FakeNFT/Core/Data/DTO/CollectionDTO.swift
