+++ FakeNFT/Core/Domain/UseCase/GetCollectionsUseCase.swift
