+++ FakeNFT/Core/Factories/CoordinatorFactory.swift
@@ -5,7 +5,7 @@ protocol CoordinatorFactory {
 	func makeOnboardingCoordinator(router: Router) -> Coordinator & OnboardingCoordinatorOutput
 	func makeAuthCoordinator(router: Router) -> Coordinator & AuthCoordinatorOutput
 
-	func makeTabbarCoordinator(container: MainDIContainer) -> (configurator: Coordinator, toPresent: Presentable?)
+	func makeTabbarCoordinator(container: MainFlowDIContainer) -> (configurator: Coordinator, toPresent: Presentable?)
 
 	func makeProfileCoordinator(navController: UINavigationController, container: ProfileFlowDIContainer) -> Coordinator
 	func makeCatalogCoordinator(navController: UINavigationController, container: CatalogFlowDIContainer) -> Coordinator
