--- FakeNFT/Core/Factories/Interface/AuthModuleFactory.swift
@@ -1,3 +1,0 @@
-protocol AuthModuleFactory {
-	func makeAuthViewController(viewModel: AuthViewModel) -> Presentable
-}
