--- FakeNFT/Core/Factories/Interface/CommonModuleFactory.swift
@@ -1,6 +1,0 @@
-protocol CommonModuleFactory {
-	func makeSplashViewController() -> Presentable
-	func makeWebViewController(viewModel: WebViewModel) -> Presentable
-	func makeErrorAlertVC(message: String, completion: (() -> Void)?) -> Presentable
-	func makeSortAlertVC<T: CustomStringConvertible>(sortCases: [T], completion: @escaping (T) -> Void) -> Presentable
-}
