--- FakeNFT/Core/Factories/Interface/OnboardingModuleFactory.swift
@@ -1,3 +1,0 @@
-protocol OnboardingModuleFactory {
-	func makeOnboardingViewController(viewModel: OnboardingViewModel) -> Presentable
-}
