--- FakeNFT/Core/Factories/Interface/ShoppingCartModuleFactory.swift
@@ -1,3 +1,0 @@
-protocol ShoppingCartModuleFactory {
-	func makeShoppinCartViewController(viewModel: ShoppingCartViewModel) -> Presentable
-}
