--- FakeNFT/Core/Factories/Interface/StatisticsModuleFactory.swift
@@ -1,3 +1,0 @@
-protocol StatisticsModuleFactory {
-	func makeStatisticsViewController(viewModel: StatisticsViewModel) -> Presentable
-}
